image_processing packaged renamed to image
