+++ lcls_tools/common/image/roi.py
