--- lcls_tools/common/image_processing/__init__.py
@@ -1,1 +1,0 @@
-from lcls_tools.common.image_processing.image import Image
