+++ lcls_tools/common/matlab2py/mat_image.py
@@ -3,7 +3,7 @@ import scipy.io as sio
 import matplotlib.pyplot as plt
 import os
 
-from lcls_tools.common.image_processing import Image
+from lcls_tools.common.image import Image
 
 
 class MatImage(object):
