+++ tests/unit_tests/lcls_tools/common/image_processing/test_image_processing.py
@@ -1,7 +1,7 @@
 import unittest
 import numpy as np
-from lcls_tools.common.image_processing.processing import ImageProcessor
-from lcls_tools.common.image_processing.roi import RectangularROI
+from lcls_tools.common.image.processing import ImageProcessor
+from lcls_tools.common.image.roi import RectangularROI
 
 
 class TestImageProcessing(unittest.TestCase):
